rename module and remove new-installation specific handling; move updates into separate module for maintainability/testability
